change: pageからtodoのリスト部分を引き抜いて再作成
